Crate theater_client

Theater Client

This module provides utilities for connecting to and interacting with a Theater server. It offers a simple, flexible interface for sending commands and receiving responses.

Modules

tcp 🔒

TCP Connection for Theater

Structs

TheaterConnection

A client connection to a Theater server

TheaterServer

Enums

ManagementCommand

ManagementError

ManagementResponse